requirements.txt
Flask-RESTful==0.3.10
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow_enum==1.5.1
flask-cors==4.0.0
